requirements.txt
matplotlib>=3.6.0
numpy<2
tqdm
praat-parselmouth>=0.4.4

